change insertDict implementation
Why: Insert dict was not properly impmlemented before. It would break
the dictionary array of sequences so that the memory could not be freed.
Now the sequence is inserted into the array without breaking it.

How: The insert function would make the table node point to the inserted
keys memory location this would break the array. Now the data contained
by the inserted key is transerfered to the sequence in the array. This
preserves the structure of the array and allows for memory to be freed.