Fix memory bug in dict class
See: n/a

Why: To free all memory at exit.

How: I changed the way that newDict is made. I adjusted pointer. Now
table points to a pointer to a sequence. and then it is made into an
array of pointers to sequences. I also changed the way that insert dict
indexes these pointers.